MDS_Video/requirements.txt
pandas
pywebhdfs
redis
rq
kafka-python
flask
gunicorn
yt-dlp==2022.6.29
